requirement.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-dotenv
python-multipart
passlib[bcrypt]
